Clean-up stale containers if the installation fails
The main reason for this changes is that UCP installer
leaves the `ucp-port-checker` containers alive if the port
check fails.

But in general, it is good to clean up anything that may have been left.
This way the `launchpad apply` cmd will be idempotent, and will not
alter the state of the machine if the installation did fail.

Signed-off-by: asankov <a.sankov96@gmail.com>